instance: Database name and Storage
New way to store an instance object into the storage.
Expose the instance object into the Config.generate_instance_dbname function.

We are not maintaing an instance list anymore and so all instances are separated
in the storage like on k8s Service Catalog.

Shared databases are still supported and we store an instance with extra
attributes like "cluster" and "database".
This permit to know if another instance is using the same database with a simple query.

By default, the Atlas Broker will return a database name based on the instance_id.